+++ src/ast/mod.rs
@@ -1,9 +1,10 @@
-pub use self::block::BlockName;
 pub use self::block::{
-    Attribute, AttributeName, AttributeValue, Block, InvalidAttributeName, InvalidAttributeValue,
-    InvalidBlockName,
+    Attribute, AttributeName, AttributeValue, Block, BlockName, InvalidAttributeName,
+    InvalidAttributeValue, InvalidBlockName,
 };
+pub use self::raw::{InvalidRaw, Raw};
 pub use self::section::Section;
 
 mod block;
+mod raw;
 mod section;