+++ src/ast/section.rs
@@ -1,12 +1,12 @@
-use std::{borrow::Cow, fmt::Display};
+use std::fmt::Display;
 
-use crate::Block;
+use crate::{Block, Raw};
 
 /// A Vue SFC section.
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub enum Section<'a> {
-    /// Represent any data before, after or between blocks.
-    Raw(Cow<'a, str>),
+    /// See [`Raw`];
+    Raw(Raw<'a>),
     /// See [`Block`].
     Block(Block<'a>),
 }
@@ -14,7 +14,7 @@ pub enum Section<'a> {
 impl Display for Section<'_> {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
-            Self::Raw(content) => content.trim_end().fmt(f),
+            Self::Raw(content) => content.fmt(f),
             Self::Block(block) => block.fmt(f),
         }
     }